+++ .credo.exs
@@ -1,0 +1,19 @@
+# .credo.exs or config/.credo.exs
+%{
+  configs: [
+    %{
+      name: "default",
+      files: %{
+        included: ["lib/", "src/", "web/", "apps/"],
+        excluded: []
+      },
+      plugins: [],
+      requires: [],
+      strict: false,
+      parse_timeout: 5000,
+      color: true,
+      checks: [
+      ]
+    }
+  ]
+}